fixing: mvd  and copy

+++ commands/dir_commands.c
@@ -1,4 +1,4 @@
-#include "../cli.h" // for linter purpose this is useless but it will make the linter "happy"
+#include "../cli.h"
 #include <stdio.h>
 #include <string.h>
 
@@ -10,7 +10,7 @@ int md() {
   }
   do {
     char name[11];
-    strncpy(name, paths, 11);
+    strncpy(name, paths, 10);
     // check if the directory name cotains a '/'
     if (strchr(name, '/') != NULL) {
       printf("md :cannot create directory recursive path");
@@ -71,63 +71,76 @@ int dir(char *paths) {
     return 1;
   Item *childrens = current_dir->dir_list.childrens;
   for (int i = 0; i < current_dir->dir_list.n_children; i++) {
+
+    if (childrens[i].attribute == 2)
+      printf("%s", "\x1b[11m");
     printf("%s ", childrens[i].name);
+    if (childrens[i].attribute == 2)
+      printf("%s", "\033[0m");
   }
   printf("\n");
+
   return 0;
 }
-int movDir(char *name, char *copyPath) {
+int mov_dir(char *name, char *copyPath) {
   int dir_idx = dir_search(name);
   if (dir_idx == -1) {
     printf("mvdir: cannot move directory '%s': No such file or directory\n",
            name);
     return 1;
   }
+  if (strchr(copyPath, '/') == NULL) {
+    printf("mvdir: invalid path\n");
+    return 1;
+  }
   // get the directory
   Item dir = current_dir->dir_list.childrens[dir_idx];
 
-  // delete the direcotry from the current_dir 
+  // delete the direcotry from the current_dir
   delete_item(dir_idx);
-  // getting the directory newName from the copyPath 
-  write_dir();
-  const char *newName = strrchr(copyPath, '/') + 1;
-  copyPath[strlen(copyPath) - strlen(newName) - 1] = '\0';
-#ifdef DEBUG
-  printf("%s", copyPath);
-#endif /* ifdef DEBUG */
+  // getting the directory newName from the copyPath
 
+  const char *newName = strrchr(copyPath, '/') + 1;
+  int copyPath_len = strlen(copyPath);
+  int newName_len = strlen(newName);
+  copyPath[copyPath_len - newName_len - 1] = '\0';
+
+  debug_print("%s", copyPath);
 
   // saving the current path and oldname
-  char current_dir_path[100];
+  char current_dir_path[100] = {0};
   strncpy(current_dir_path, current_dir->path, 99);
-  char oldname[11];
-  strncpy(oldname,dir.name,10);
+  char oldname[11] = {0};
+  strncpy(oldname, dir.name, 10);
 
   copyPath = strtok(copyPath, "/");
-
-  strncpy(dir.name, newName, 10);
 
   while (copyPath != NULL) {
     if (change_dir(copyPath) /* the file doesn't exist */) {
       // reset the directory (back to the root)
       free_current_dir();
       dir_init();
-      // go to the current path 
+      // go to the current path
       copyPath = strtok(current_dir_path, "/");
       while (copyPath != NULL) {
         change_dir(copyPath);
         copyPath = strtok(NULL, "/");
       }
       // add the directory again
-      strncpy(dir.name,oldname,10);
+      strncpy(dir.name, oldname, 10);
       add_to_dir(dir);
       write_dir();
       return 1;
     }
     copyPath = strtok(NULL, "/");
   }
-  add_to_dir(dir);
-  write_dir();
+  if (dir_search(newName) == -1) {
+    strncpy(dir.name, newName, 10);
+    add_to_dir(dir);
+    write_dir();
+  } else {
+    printf("mvdir: cannot move directory '%s': File exists\n", newName);
+  }
   free_current_dir();
   dir_init();
   copyPath = strtok(current_dir_path, "/");
@@ -135,5 +148,6 @@ int movDir(char *name, char *copyPath) {
     change_dir(copyPath);
     copyPath = strtok(NULL, "/");
   }
+  write_dir();
   return 0;
 }
